refactor: move link to ui and change default styling
